Support CREATE TABLE statements w/ IdentifierLists in them. (#48)
It looks like #47 is failing because the code we use to parse SQL files to detect what tables are being created by them isn't accounting for `IdentifierList` tokens from the `sqlparse` library, which apparently the SQL for WoW's new `create_bldgs_table` uses.  This adds support for them and makes sure that both WoW's new and old `create_bldgs_table` parse as expected.