On push to master, build and push production container to Docker Hub (#51)
This is an attempt to use Docker multi-stage builds to both test the code and push the production container to Docker Hub. This ensures that once CircleCI says a build is done, we know that our production infrastructure is updated, rather than having to wait for Docker Hub's auto-builder to complete (and possibly hiccup and fail, which has happened in the past).