Generate WoW portfolio graphs. (#81)
Oy. #79 didn't do what we wanted--it only made an empty table for WoW portfolio graphs--because nycdb-k8s-loader doesn't actually run any of WoW's python code. It only goes through WoW's who-owns-what.yml, extracts its list of SQL files and runs them. This was done back in the day because WoW itself isn't really a Python package the way that nycdb is (and turning it into one isn't particularly easy since it includes lots of dependencies that nycdb-k8s-loader doesn't actually need, like Django).

So this is a terrible hack that installs just enough dependencies (networkx) for WoW's portfoliograph package to run, and then uses it to populate the WoW portfolio graphs table.

It is not awesome. But it should hopefully make nycdb-k8s-loader actually populate the portfolio graphs table.