Update WOW_REV to incorporate new tax exemption data and `wow_bldgs` sql (#84)
This PR updates our reference to the Who Owns What codebase so that we can reference the updated SQL files in JustFixNYC/who-owns-what#550 in our auto-updating instance of nycdb.

In doing so, this PR caught a small issue with our testing infrastructure where the inferred table name from WOW for the create buildings table was detected as wow_bldgs_temporary, missing the line of SQL that changes the name of this temporary table back to wow_bldgs at the end of the file. So, this PR also updates the parse_created_tables function to additionally detect the keywords ALTER TABLE .... RENAME TO in a SQL file and update the extracted names of created tables accordingly in the output. I added some extra testing infrastructure for this case as well.