add wow schema to search path when in temp schema (#116)
This PR adds wow to the search_path used when in the temp schema. This is because when creating wow_bldgs we need access to wow.oca_evictions_bldgs. I think this option is easier than to put the explicit wow. prefix in the query because that complicates everything is testing where all the schemas don't work the same way.