update for nycdb updates to package setup and psycopg version (#130)
This PR updates to the latest version of NYCDB. More than the usual updating of the commit hash and new datasets, there were a few additional changes required for the Dockerfile.

In nycdb/nycdb#257 the python package setup files were changed - removing the setup.py and instead using pyproject.toml. Apparently without a setup.py you can't install in editable mode (pip install -e .), so that was removed, but isn't important. Also in that nycdb PR the new psycopg (3) replaces psycopg2, and this caused some issues and conflicts with packages, so required a few more changes - I've updated the dockerfile to use python 3.9, upgraded the pip version, and bumped the versions of a couple packages. And changing to python 3.9 required upgrading this repo's psycopg2 version, and that required a couple minor code changes.

With this PR we're adding a few new datasets, but most use the default annual schedule so don't appear in the code. They are:

dof_tax_lien_sale_list
pluto_23v1
vacate_orders
mci_applications