update nycdb commit hash (#132)
updating the nycdb version for this branch where I fixed an issue with our db inserts that was slowing everything down a ton. It's not merged into main yet, but I've tested it out locally with k8s-loader, and wanted to get this up before it does another round of super slow updates.