update nycdb hash for testing revert to psycopg2 (#133)
Since the update to psycopg3 in nycdb (nycdb/nycdb#257) we've been having some problems with the auto-updating. First it was an issue with the method of inserting rows of data which slowed everything way down so jobs were either taking forever or stalling out and breaking. That seems to have been fixed by using multi-row insert statements (nycdb/nycdb#266), and now the jobs take the normal amount of time.

However, there remains something weird going on when trying to run the new pluto (22v1, 23v1) datasets. they work fine with just nycdb locally, also work running k8s-loader jobs locally with a local db and a remote db, but when the exact same jobs run on kubernetes they just stall out after having created the table just before the rows are inserted. I can't think of what the issue is, but want to try reverting back to psycpg2 temporarily and try that.