update wow, fix tests, remove old ssh, add aws keys (#136)
We recently made some changes for the OCA address-level data (see JustFixNYC/who-owns-what#812) and so need to make some small changes here as well. We no longer need to securely store the SSH key in kubernetes, and instead use AWS access keys to get the data as csv from s3.