update nycdb version for new oca_metadata table (#144)
We have created a new table to add to the existing oca dataset of housing court records. This isn't directly in the raw data they provide, but records some important metadata about when records appear in OCA's extracts, which will help us to more accurately track the lagged reporting of cases that we know is an issue, as well as have a sense of the number of records they request us to delete (sealed records, etc.).