update nycdb version for test of switching to copy import method (#149)
There is a proposed change to nycdb to use the copy method for inserting data into the db (nycdb/nycdb#328), and this appears to considerably speed up the process. it's still a work in progress, but I want to temporarily update the docker image to test this on kubernetes and see what impact there is on cpu usage since that continues to be an issue.