add k8s cronjob resource requests values (mem, cpu) (#151)
This edits the kubernetes cronjob template yaml definition to include default resource request limits for memory and cpu. With these resource request values kubernetes ensures that before the scheduled cronjob begins (pod created) there is at least this much memory and cpu available for the life of the job. In practice this means that it will delay the start of the job from the cronjob time (since we have a lot all scheduled for the same times) but we will avoid the problem of exceeding our resources and crashing.