update wow for latest signature changes (#161)
Update for latest version of wow for signature backend updates.

Got a weird error from mypy in CI tests and it said to update. I bumped us up to the latest version, and added some installs for types on some packages and ignored one weird type error in the code (the suggested type was some massive mashup of types that seems wrong).