use NYC local time in k8s cron job scheduling (#173)
Specify the timezone in our k8s cron job definitions so that it's easier to read on dashboard, in scheduling script, and we also don't have to worry about things shifting around with daylight savings time

I tested this out by just updating a single job and it was triggered at the expected time. I'll now rebuild all the nycdb jobs and re-upload them.