use alternating days for jobs (#177)
For all the jobs that we are using either every day, every other day, or yearly. And we tend to put a lot of jobs on every-other day (especially big ones like ACRIS and ones we don't use in our products). But we are using a single crontab for this that has every job running on odd days only. Ass the list of jobs has grown, this has lead to those days running really long (past our plan to have WOW updated before the work day). So in this PR we change to odd days and even days and put ACRIS on its own set of days (we might need to rebalance later, but this seems to help for now.)

Also, I've changed to have all the nightly jobs start a little earlier (11pm instead of midnight) to give us more time to finish in time for our morning jobs (wow, signature, etc.) so that everything is done before the work day (heavy traffic hours)

Also, I added a new nycdb job for DOF Property Valuation and Assessments