use backup schema for wow tables (#194)
For the our WOW tables created via nycdb-k8s-loader, there has always been a
brief period when the tables are recreated that they are inaccessible causing
"table not found" errors on our tools. As a solution to this we are creating
a new schema backup where we will move all the wow tables before moving the
new versions from their temporary schema to the wow schema. This way we are
never deleting the only copy of wow tables - there will always be another
copy in the backup schema. So we'll just change our wow-role search path to
include backup and so if it doesn't find the tables in wow it will find
them in backup until the new versions are ready again in wow

Once we confirm that this solves our problem with the downtime for wow tables used in wow apis, we can consider doing the same approach for other jobs. So for now I've only done this for the wow job, not the nycdb or others, but we could do the same for those if we want. If we do expand the use of it though I think we should add a step to delete the backup tables after the new ones are ready, so we aren't permanently keeping duplicates of every table.

I've tested this out on a temporary clone of our nycdb database and it seems to be working just fine.

Commands to set up backup schema and updating permissions