update nycdb for pluto_latest_districts, handle missing nycdb schema (#199)
updates nycdb for this yet-to-be-merged PR nycdb/nycdb#386 which adds pluto_latest_districts and pluto_latests_districts_25a. There are still some changes that will need to be made on the nycdb side to make this dataset without source data pattern work well generally, but for now it's working to get it up in the DB and allow for the wow dataset to depend on this new nycdb dataset.