update url for geosupport download in dockerfile (#207)
A new version of Geosupport has been released, 25a, and this updates the dockerfile for the new download file. This also solves a confusing build error we were getting suddenly on CI. It's good to keep this updated regularly so that it keeps pace with the version of PLUTO we get in pluto_latest.