update nycdb for zipcodes dedupe and districts table (#210)
Updates NYCDB version to include recent PRs:

remove duplicate rows for zipcodes by joining multiple areas of the same zipcode into a single multipolygon geom and only one row per zipcode combine multiple rows per zipcode into single multiploygon record nycdb/nycdb#388
add pluto_latest_districts (and 25a version), which was previously only on a branch but now is improved and merged into main add SQL-only dataset with dependencies for pluto_latest_districts nycdb/nycdb#386
I haven't added the new pluto_latest_districts datasets to the schedule or k8s cluster since we don't yet have a way to know when the updates are required. For now we can just run manually when pluto_latest updates, but later we might consider added to the update/change tracking system for these datasets without source data to instead check the update dates of the dependency datasets, so that the dataset will only be recreated when with pluto or the boundaries dataset updates.

I also made a small change to a util function that gets all tables for an nycdb dataset. Since some datasets create new tables in the associated SQL script it looks through for "create table", but now with the new nycdb system for listing derived tables it was getting duplicates of zipcode since it's listed in the yaml then crated in the sql.