revert nycdb version after hpd reg issue resolved (#222)
The hpd reg files on open data were published with an error in october, so we needed to download archived versions of the files, upload those to s3, make a temp branch on nycdb to change the sources files, and then user that temp nycdb version here. HPD has now corrected the problem and published update data, so we can revert that change.